move code to parse Package from cargo_metadata
